app/requirements/common.txt
newsapi-python==0.2.6
pycountry==20.7.3
requests==2.24.0
python-dateutil==2.8.1
folium==0.11.0
geopandas==0.8.1
numpy==1.18.5
pandas==1.0.5
psycopg2==2.8.5
APScheduler==3.6.3
django-storages==1.9.1
boto3==1.14.25
dj_database_url==0.5.0
django_heroku==0.3.1
better-exceptions==0.2.2
gunicorn==20.0.4
docutils==0.15.2
matplotlib==3.3.0
descartes==1.1.0
whitenoise==5.1.0